# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /registros:
    get:
      operationId: listRegistros
      tags: [registros]
      summary: List meeting records (leader/auxiliary)
      responses:
        "200":
          description: Meeting records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegistroSummary"
    post:
      operationId: createRegistro
      tags: [registros]
      summary: Create a meeting record (leader publishes directly; auxiliary creates pending)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistroInput"
      responses:
        "201":
          description: Meeting record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistroDetail"
        "400":
          description: Invalid input (e.g. donations without an active campaign)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /registros/atividades:
    get:
      operationId: listRegistroActivities
      tags: [registros]
      summary: List the activity catalog (seeds the builtin activities on first use)
      responses:
        "200":
          description: Activity catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RegistroActivity"
    post:
      operationId: createRegistroActivity
      tags: [registros]
      summary: Create a custom catalog activity (leader/auxiliary)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistroActivityInput"
      responses:
        "201":
          description: Activity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistroActivity"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /registros/atividades/{id}:
    delete:
      operationId: deleteRegistroActivity
      tags: [registros]
      summary: Remove a custom catalog activity (builtin activities cannot be removed)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "400":
          description: Builtin activity cannot be removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /registros/{id}:
    get:
      operationId: getRegistro
      tags: [registros]
      summary: Get a meeting record with attendance, activities, donations and album
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting record detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistroDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateRegistro
      tags: [registros]
      summary: Update a meeting record (auxiliary edits send published records back to pending)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistroUpdate"
      responses:
        "200":
          description: Updated meeting record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistroDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteRegistro
      tags: [registros]
      summary: Delete a meeting record (leader only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResult"
        "403":
          description: Leader only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /registros/{id}/approve:
    post:
      operationId: approveRegistro
      tags: [registros]
      summary: Approve a pending meeting record (leader only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting record published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistroDetail"
        "403":
          description: Leader only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    OkResult:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
...
    RegistroSummary:
      type: object
      properties:
        id:
          type: integer
        seq:
          type: integer
          description: Sequential meeting number
        eventDate:
          type: string
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
        status:
          type: string
          enum: [pending, published]
        presentCount:
          type: integer
        createdByName:
          type: ["string", "null"]
      required:
        - id
        - seq
        - eventDate
        - status
        - presentCount
        - createdByName
    RegistroAttendee:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [member, guest]
      required:
        - userId
        - name
        - status
    RegistroActivityEntry:
      type: object
      properties:
        id:
          type: integer
        atividadeId:
          type: ["integer", "null"]
          description: Catalog activity id, null when the catalog entry was removed
        name:
          type: string
          description: Snapshot of the activity name
        responsavelId:
          type: ["integer", "null"]
        responsavelName:
          type: ["string", "null"]
        durationMin:
          type: ["integer", "null"]
      required:
        - id
        - atividadeId
        - name
        - responsavelId
        - responsavelName
        - durationMin
    RegistroDonationItem:
      type: object
      properties:
        id:
          type: integer
        itemName:
          type: string
        quantity:
          type: integer
      required:
        - id
        - itemName
        - quantity
    RegistroAlbum:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        driveUrl:
          type: ["string", "null"]
      required:
        - id
        - title
        - driveUrl
    RegistroDetail:
      type: object
      properties:
        id:
          type: integer
        seq:
          type: integer
        eventDate:
          type: string
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
        status:
          type: string
          enum: [pending, published]
        notes:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        approvedByName:
          type: ["string", "null"]
        presentes:
          type: array
          items:
            $ref: "#/components/schemas/RegistroAttendee"
        atividades:
          type: array
          items:
            $ref: "#/components/schemas/RegistroActivityEntry"
        arrecadacao:
          type: array
          items:
            $ref: "#/components/schemas/RegistroDonationItem"
        album:
          oneOf:
            - $ref: "#/components/schemas/RegistroAlbum"
            - type: "null"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - seq
        - eventDate
        - status
        - notes
        - createdByName
        - approvedByName
        - presentes
        - atividades
        - arrecadacao
        - album
        - createdAt
        - updatedAt
    RegistroNewGuest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
      required:
        - name
    RegistroActivityInputEntry:
      type: object
      properties:
        atividadeId:
          type: integer
          description: Catalog activity id, when picked from the catalog
        name:
          type: string
          minLength: 1
          description: Activity name snapshot
        responsavelId:
          type: integer
        durationMin:
          type: integer
          minimum: 1
      required:
        - name
    RegistroDonationInputEntry:
      type: object
      properties:
        item:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 1
      required:
        - item
        - quantity
    RegistroInput:
      type: object
      properties:
        eventDate:
          type: string
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
        presentes:
          type: array
          description: Ids of members/guests present
          items:
            type: integer
        novosConvidados:
          type: array
          description: New guests to create (status guest, no tags) and mark present
          items:
            $ref: "#/components/schemas/RegistroNewGuest"
        atividades:
          type: array
          items:
            $ref: "#/components/schemas/RegistroActivityInputEntry"
        albumId:
          type: integer
        arrecadacao:
          type: array
          description: Donated items, linked to the currently active campaign
          items:
            $ref: "#/components/schemas/RegistroDonationInputEntry"
        notes:
          type: string
      required:
        - eventDate
        - presentes
    RegistroUpdate:
      type: object
      properties:
        eventDate:
          type: string
          pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
        presentes:
          type: array
          description: Replaces the attendance list when provided
          items:
            type: integer
        novosConvidados:
          type: array
          items:
            $ref: "#/components/schemas/RegistroNewGuest"
        atividades:
          type: array
          description: Replaces the activity list when provided
          items:
            $ref: "#/components/schemas/RegistroActivityInputEntry"
        albumId:
          type: ["integer", "null"]
        arrecadacao:
          type: array
          description: Replaces the donation entries of this record when provided
          items:
            $ref: "#/components/schemas/RegistroDonationInputEntry"
        notes:
          type: ["string", "null"]
    RegistroActivity:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        hasDuration:
          type: boolean
        builtin:
          type: boolean
      required:
        - id
        - name
        - hasDuration
        - builtin
    RegistroActivityInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        hasDuration:
          type: boolean
          default: true
      required:
        - name
...