# lib/api-zod/src/generated/types/registroActivity.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegistroActivity {
  id: number;
  name: string;
  hasDuration: boolean;
  builtin: boolean;
}